# entc/gen: improve assets formatting performance