@a8m 2023-01-13
Thanks for sending this patch, @crossworth.

I actually tried this more than two years ago, and sent a patch to goimports to support this (https://github.com/golang/tools/commit/1f7a813c0df1b2a1c1526e54c5ee4aa940c58d14), but from what I remember, the results were not as good for FB internal projects - this was due to the way goimports works (I can elaborate if that’s needed).

However, Ent was improved a lot since then (goimports needs to unimport mainly), so I don’t mind giving it another try. #3235 feels too slow for me. We either miss some imports there, or it’s slower because `go generate` contains more commands.